requirements.txt
paho-mqtt==2.1.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
asyncio==4.0.0
fastapi==0.115.12
uvicorn==0.34.0